feat(wire): per-direction u64 packet number (WIRE_VERSION 3) — PhantomUDP P4.0

+++ core/benches/transport_bench.rs
@@ -220,7 +220,7 @@ fn bench_encryption(c: &mut Criterion) {
             b.iter_batched(
                 || {
                     let seq = encrypt_seq.fetch_add(1, Ordering::Relaxed);
-                    PacketHeader::new(session_id, 1, seq, flags)
+                    PacketHeader::new(session_id, 1, seq as u64, flags)
                 },
                 |header| {
                     let encrypted = server_session.encrypt_packet(&header, &data);
@@ -236,7 +236,7 @@ fn bench_encryption(c: &mut Criterion) {
             b.iter_batched(
                 || {
                     let seq = decrypt_seq.fetch_add(1, Ordering::Relaxed);
-                    let header = PacketHeader::new(session_id, 2, seq, flags);
+                    let header = PacketHeader::new(session_id, 2, seq as u64, flags);
                     let encrypted = server_session
                         .encrypt_packet(&header, &data)
                         .expect("encrypt setup");
@@ -301,7 +301,7 @@ fn bench_throughput(c: &mut Criterion) {
         b.iter_batched(
             || {
                 let seq = seq_counter.fetch_add(1, Ordering::Relaxed);
-                PacketHeader::new(session_id, 1, seq, flags)
+                PacketHeader::new(session_id, 1, seq as u64, flags)
             },
             |header| {
                 let encrypted = server_session.encrypt_packet(&header, &data).unwrap();
